cleanup and task moved to api folder
