+++ Tests/StorageServiceTest.cpp
@@ -3,7 +3,7 @@
 //
 
 
-#include<gtest/gtest.h>
+#include <gtest/gtest.h>
 
 #include "Model/StorageService.h"
