+++ Tests/TaskDtoTest.cpp
@@ -6,7 +6,7 @@
 // Created by Ilya on 7/26/2020.
 //
 
-#include<gtest/gtest.h>
+#include <gtest/gtest.h>
 
 #include "API/TaskDTO.h"
 using testing::Eq;
