+++ Tests/TaskEntityTest.cpp
@@ -2,7 +2,7 @@
 // Created by Ilya on 7/26/2020.
 //
 
-#include<gtest/gtest.h>
+#include <gtest/gtest.h>
 
 #include "Model/TaskEntity.h"
 using testing::Eq;
