+++ Tests/TaskTest.cpp
@@ -2,9 +2,9 @@
 // Created by Ilya on 7/26/2020.
 //
 
-#include<gtest/gtest.h>
+#include <gtest/gtest.h>
 
-#include "Model/Task.h"
+#include "API/Task.h"
 
 using testing::Eq;
