+++ src/Model/HashTaskID.h
@@ -5,7 +5,7 @@
 #ifndef TODOLIST_HASHTASKENTITY_H
 #define TODOLIST_HASHTASKENTITY_H
 
-#include"TaskEntity.h"
+#include "TaskEntity.h"
 
 class HashTaskID {
 public:
